# Conflicts:
#	src/main/java/ua/lviv/iot/repository/UserRepository.java
#	src/main/java/ua/lviv/iot/service/AdminService.java

+++ src/main/java/ua/lviv/iot/repository/CandidateRepository.java
@@ -5,5 +5,5 @@ import org.springframework.stereotype.Repository;
 import ua.lviv.iot.model.election.candidate.Candidate;
 
 @Repository
-public interface CandidateRepository extends JpaRepository<Candidate, Long> {
+public interface CandidateRepository extends JpaRepository<Candidate, Integer> {
 }
